test(ebpf): 稳定统计断言并补充 SKB 模式说明
- 移除 syn_cookie 中多余的 bucket as u64 转换（eBPF clippy -D warnings）
- netns Test 3/13 改为轮询 stats API，等待用户态 1s 同步 eBPF 计数，
  避免 sleep 0.5s 读到全零快照导致误判
- AGENTS 记录 ESHIELD_XDP_MODE=skb / ESHIELD_TEST_XDP_MODE=skb 用法，
  以及内核 7.0 verifier 约束下 PORT_ACL=32、L7=8 的上限
- 验证：SKB 模式 netns 全部 PASS（含 conn_track）、Hub-Node 8 项 PASS、
  full_attack 冒烟 PASS；原生模式黑名单/统计验证 PASS（不含 SYN Cookie
  XDP_TX 场景）

+++ eshield-ebpf/src/syn_cookie.rs
@@ -154,7 +154,7 @@ fn syn_challenge(
 
     let cookie = build_cookie_runtime(
         (saddr as u64) | ((daddr as u64) << 32),
-        (sport as u64) | ((dport as u64) << 16) | ((bucket as u64) << 32),
+        (sport as u64) | ((dport as u64) << 16) | (bucket << 32),
         secret_bytes,
         mss_idx,
     );
@@ -223,7 +223,7 @@ pub fn handle_ack(pc: &PacketCtxRef, ip: *const IpHdr, tcp: *const TcpHdr) -> u3
 
         let computed = build_cookie_runtime(
             (saddr as u64) | ((daddr as u64) << 32),
-            (sport as u64) | ((dport as u64) << 16) | ((bucket as u64) << 32),
+            (sport as u64) | ((dport as u64) << 16) | (bucket << 32),
             secret_bytes,
             mss_idx,
         );
